Edited Code for Use with different lift
The build team went and did it: they changed the system we are using
for the lift mechanism after I had written the code for it. We are
using a CIM Motor now as a sort of winch, instead of the actuator.
Still have to mount the encoder, but should be able to test tomorrow